+++ hal/sensors/inc/reflectanceSensor.hpp
@@ -18,13 +18,13 @@ private:
     uint8_t count;
     iPin * power;
     iTimer<uint16_t> * timer;
-    uint8_t * readings;
+    uint16_t * readings;
     iPin * sensor[];
 
 //functions
 public:
     ~ReflectanceSensor();
-    ReflectanceSensor(iPin * sensor[], uint8_t count, iPin * power, iTimer<uint16_t> * timer, uint8_t * readings);
+    ReflectanceSensor(iPin * sensor[], uint8_t count, iPin * power, iTimer<uint16_t> * timer, uint16_t * readings);
     SensorData read();
 protected:
 private:
